requirements.txt
keras==2.12.0
matplotlib==3.6.2
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.8.0.76
Pillow==9.3.0
Pillow==10.2.0
scikit_learn==1.1.3
SQLAlchemy==2.0.25
tensorflow==2.12.0
tensorflow_intel==2.12.0
torch==2.1.2+cu118
torchvision==0.16.2+cu118
